--- tests/test_Action_abstractclass.py
@@ -1,4 +1,0 @@
-from pytraj.actions.Action import Action
-
-a = Action()
-a.do_action()
